Move collisions to a separate crate
