+++ collisions/src/prelude.rs
@@ -1,0 +1,4 @@
+pub use crate::colliders::{Collider, ColliderData};
+pub use crate::plugin::{ColliderBundle, ColliderComponent, ColliderDespawnable};
+pub use crate::shapes::{ColliderShape, ColliderShapeType};
+pub use crate::store::{ColliderIdResolver, ColliderStore, ALL_LAYERS};
