+++ common/Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+"bevy" = { version = "0.14"}
+rand = "0.8.5"
