+++ vs-rs/src/player.rs
@@ -1,16 +1,14 @@
 use crate::assets::GameAssets;
-use crate::collisions::colliders::ColliderData;
-use crate::collisions::plugin::{ColliderBundle, ColliderComponent};
-use crate::collisions::shapes::ColliderShapeType;
-use crate::collisions::store::{ColliderIdResolver, ColliderStore};
 use crate::enemy::Enemy;
 use crate::input::PlayerControls;
 use crate::movement::behaviors::SteerSeek;
-use crate::movement::{PhysicsParams, Position, SteeringBundle, SteeringHost};
+use crate::movement::{PhysicsParams, SteeringBundle, SteeringHost};
 use crate::stats::*;
 use crate::AppState;
 use bevy::sprite::Anchor;
 use bevy::{input::gamepad::GamepadSettings, prelude::*};
+use collisions::prelude::*;
+use common::Position;
 use std::time::Duration;
 
 pub struct PlayerPlugin;
