+++ vs-rs/src/worlds/worldgen/room.rs
@@ -1,7 +1,8 @@
-use crate::{assets::rooms::MapAsset, collisions::Rect};
+use crate::assets::rooms::MapAsset;
 use bevy::prelude::*;
+use common::FRect;
 
 pub struct WorldRoom {
     pub map_asset: Handle<MapAsset>,
-    pub rect: Rect,
+    pub rect: FRect,
 }
